Catch bad dates parsing in value_types
Return False when evaluating equality of an invalid Wikibase date.

Given that Wikibase doesn't ensure valid dates we must allow for
the possibility that we will be unable to parse them.

This patch specifically adresses invalid Wikibase day precision values.

It follows the existing pattern for failing to parse malformed foreign
values by not throwing and returning False.

It does not consider attempting to parse non-4 digit years

Bug: T254587